pyproject.toml
[project]
name = "stock-trading"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py",
    "pandas-ta",
    "apscheduler",
    "loguru",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
